updated item-service

+++ src/main/java/com/demo/item_service/dao/entity/ItemEntity.java
@@ -30,7 +30,7 @@ public class ItemEntity {
 	private int catalogueId;
 	
 	@Column(name="tailor_id")
-	private int tailorId;
+	private long tailorId;
 	
 	@Column(name="item_status")
 	private String itemStatus;
